--- src/main/java/Seoul_Milk/sm_server/domain/taxInvoiceValidationHistory/service/TaxInvoiceValidationHistoryService.java
@@ -1,20 +1,0 @@
-package Seoul_Milk.sm_server.domain.taxInvoiceValidationHistory.service;
-
-import Seoul_Milk.sm_server.domain.taxInvoice.enums.ProcessStatus;
-import Seoul_Milk.sm_server.domain.taxInvoiceValidationHistory.dto.TaxInvoiceValidationHistoryResponseDTO;
-import Seoul_Milk.sm_server.domain.taxInvoiceValidationHistory.dto.TaxInvoiceValidationHistoryResponseDTO.GetModalResponse;
-import Seoul_Milk.sm_server.domain.taxInvoiceValidationHistory.dto.TaxInvoiceValidationHistoryRequestDTO.ChangeTaxInvoiceRequest;
-import Seoul_Milk.sm_server.domain.taxInvoiceValidationHistory.dto.TaxInvoiceValidationHistoryRequestDTO.TaxInvoiceRequest;
-import Seoul_Milk.sm_server.login.entity.MemberEntity;
-
-public interface TaxInvoiceValidationHistoryService {
-    TaxInvoiceValidationHistoryResponseDTO.TaxInvoiceSearchResult searchByProviderOrConsumer(MemberEntity memberEntity, ProcessStatus processStatus, String poc, int page, int size);
-
-    Void deleteValidationTaxInvoice(MemberEntity memberEntity, TaxInvoiceRequest taxInvoiceRequest);
-
-    Void tempSave(MemberEntity memberEntity, TaxInvoiceRequest taxInvoiceRequest);
-
-    GetModalResponse showModal(Long taxInvoiceId);
-
-    Void changeColunm(MemberEntity memberEntity, ChangeTaxInvoiceRequest changeTaxInvoiceRequest);
-}
